+++ example/lib/main.dart
@@ -1,7 +1,9 @@
-import 'package:flutter/material.dart';
 import 'dart:async';
+import 'dart:typed_data';
 
-import 'package:flutter/services.dart';
+import 'package:flutter/material.dart';
+import 'package:image_picker/image_picker.dart';
+import 'package:permission_handler/permission_handler.dart';
 import 'package:wallpaper_handler/wallpaper_handler.dart';
 
 void main() {
@@ -16,55 +18,93 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String _platformVersion = 'Unknown';
-  final _wallpaperHandlerPlugin = WallpaperHandler.instance;
+  bool _isSetLoading = false;
+  bool _isGetLoading = false;
 
-  @override
-  void initState() {
-    super.initState();
-    initPlatformState();
-  }
-
-  // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> initPlatformState() async {
-    String platformVersion;
-    // Platform messages may fail, so we use a try/catch PlatformException.
-    // We also handle the message potentially returning null.
-    try {
-      platformVersion = await _wallpaperHandlerPlugin.getPlatformVersion() ??
-          'Unknown platform version';
-
-      const assetPath = 'assets/images/test.jpg';
-      const wallpaperLocation = WallpaperLocation.homeScreen;
-      const cropBounds = Rect.fromLTRB(100, 100, 200, 100);
-      final bool isWallpaperSetSuccessfully = await WallpaperHandler.instance
-          .setWallpaperFromAssetWithCrop(
-              assetPath, wallpaperLocation, cropBounds);
-    } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
-    }
-
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
-    if (!mounted) return;
-
-    setState(() {
-      _platformVersion = platformVersion;
-    });
-  }
+  Uint8List? _imageBytes;
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('Plugin example app'),
+          title: const Text('Wallpaper Handler Example'),
         ),
         body: Center(
-          child: Text('Running on: $_platformVersion\n'),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              _isSetLoading
+                  ? const CircularProgressIndicator()
+                  : ElevatedButton(
+                      onPressed: () {
+                        _setWallpaper(context);
+                      },
+                      child: const Text('Set Wallpaper'),
+                    ),
+              const SizedBox(height: 24),
+              _isGetLoading
+                  ? const CircularProgressIndicator()
+                  : ElevatedButton(
+                      onPressed: () {
+                        _getWallpaper();
+                      },
+                      child: const Text('Get Wallpaper'),
+                    ),
+              const SizedBox(height: 24),
+              _imageBytes != null
+                  ? Image.memory(
+                      _imageBytes!,
+                      width: 300,
+                      height: 300,
+                    )
+                  : const SizedBox.shrink(),
+            ],
+          ),
         ),
       ),
     );
+  }
+
+  Future<void> _setWallpaper(BuildContext context) async {
+    final ImagePicker picker = ImagePicker();
+    final XFile? image = await picker.pickImage(source: ImageSource.gallery);
+    setState(() {
+      _isSetLoading = true;
+    });
+    if (image != null) {
+      _imageBytes = null;
+      const wallpaperLocation = WallpaperLocation.homeScreen;
+      final bool isWallpaperSetSuccessfully = await WallpaperHandler.instance.setWallpaperFromFile(
+        image.path,
+        wallpaperLocation,
+      );
+      if (isWallpaperSetSuccessfully) {
+        if (context.mounted) {
+          ScaffoldMessenger.of(context).showSnackBar(const SnackBar(content: Text('Wallpaper set successfully.')));
+        }
+      }
+      setState(() {
+        _isSetLoading = false;
+      });
+    } else {
+      setState(() {
+        _isSetLoading = false;
+      });
+    }
+  }
+
+  Future<void> _getWallpaper() async {
+    if ((await Permission.manageExternalStorage.request()).isGranted) {
+      setState(() {
+        _isGetLoading = true;
+      });
+      const wallpaperLocation = WallpaperLocation.homeScreen;
+      _imageBytes = await WallpaperHandler.instance.getWallpaper(wallpaperLocation);
+
+      setState(() {
+        _isGetLoading = false;
+      });
+    }
   }
 }