+++ test/wallpaper_handler_test.dart
@@ -1,7 +1,7 @@
+import 'dart:typed_data';
 import 'dart:ui';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:wallpaper_handler/src/enums.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 import 'package:wallpaper_handler/src/wallpaper_handler_method_channel.dart';
 import 'package:wallpaper_handler/src/wallpaper_handler_platform_interface.dart';
@@ -14,30 +14,24 @@ class MockWallpaperHandlerPlatform
   Future<String?> getPlatformVersion() => Future.value('42');
 
   @override
-  Future<bool> setWallpaperFromFile(
-      String filePath, WallpaperLocation wallpaperLocation) {
-    // TODO: implement setWallpaperFromFile
+  Future<Uint8List?> getWallpaper(WallpaperLocation wallpaperLocation) {
+    // TODO: implement getWallpaper
     throw UnimplementedError();
   }
 
   @override
   Future<bool> setWallpaperFromAsset(
-      String assetPath, WallpaperLocation wallpaperLocation) {
+      String assetPath, WallpaperLocation wallpaperLocation,
+      {Rect? cropBounds}) {
     // TODO: implement setWallpaperFromAsset
     throw UnimplementedError();
   }
 
   @override
-  Future<bool> setWallpaperFromAssetWithCrop(
-      String assetPath, WallpaperLocation wallpaperLocation, Rect cropBounds) {
-    // TODO: implement setWallpaperFromAssetWithCrop
-    throw UnimplementedError();
-  }
-
-  @override
-  Future<bool> setWallpaperFromFileWithCrop(
-      String filePath, WallpaperLocation wallpaperLocation, Rect cropBounds) {
-    // TODO: implement setWallpaperFromFileWithCrop
+  Future<bool> setWallpaperFromFile(
+      String filePath, WallpaperLocation wallpaperLocation,
+      {Rect? cropBounds}) {
+    // TODO: implement setWallpaperFromFile
     throw UnimplementedError();
   }
 }